refactor: Renamed AdLibPieceService to HttpAdLibPieceService and the abstraction to AdLibPieceService.
