+++ src/app/core/core.module.ts
@@ -5,7 +5,7 @@ import { HttpErrorService } from './services/http-error.service'
 import { HttpClientModule } from '@angular/common/http'
 import { HttpBasicRundownService } from './services/http-basic-rundown.service'
 import { HttpRundownService } from './services/http-rundown.service'
-import { AdLibPieceService } from './services/ad-lib-piece.service'
+import { HttpAdLibPieceService } from './services/http-ad-lib-piece.service'
 import { RundownEventObserver } from './services/events/rundown-event-observer.service'
 import { RundownStateService } from './services/rundown-state.service'
 import { RundownEventParser } from './services/events/rundown-event-parser.interface'
@@ -15,6 +15,7 @@ import { EntityParser } from './services/entity-parser.interface'
 import { ZodEntityParser } from './services/zod-entity-parser.service'
 import { ZodRundownEventParser } from './services/events/zod-rundown-event-parser.service'
 import { EventSystemModule } from '../event-system/event-system.module'
+import { AdLibPieceService } from './services/ad-lib-piece-service.interface'
 
 @NgModule({
   declarations: [],
@@ -28,11 +29,12 @@ import { EventSystemModule } from '../event-system/event-system.module'
     HttpClientModule,
     HttpBasicRundownService,
     HttpRundownService,
-    AdLibPieceService,
+    HttpAdLibPieceService,
     RundownEventObserver,
     { provide: BasicRundownService, useClass: HttpBasicRundownService },
     RundownStateService,
     BasicRundownStateService,
+    { provide: AdLibPieceService, useClass: HttpAdLibPieceService },
     { provide: RundownEventParser, useClass: ZodRundownEventParser },
     { provide: EntityParser, useClass: ZodEntityParser },
   ]
