--- src/app/core/interfaces/ad-lib-piece-service-interface.ts
@@ -1,9 +1,0 @@
-import {Observable} from 'rxjs';
-import {Identifier} from '../models/identifier';
-
-export interface AdLibPieceServiceInterface {
-
-  fetchAdLibPieceIdentifiers(rundownId: string): Observable<Identifier[]>
-
-  executeAdLibPiece(rundownId: string, adLibIdentifier: Identifier): Observable<void>
-}
