+++ src/app/core/services/events/rundown-event-observer.service.spec.ts
@@ -1,6 +1,6 @@
 import { RundownEventObserver } from './rundown-event-observer.service'
 import { anyString, anything, instance, mock, verify, when } from '@typestrong/ts-mockito'
-import { EventObserver } from './event-observer.interface'
+import { EventObserver } from '../../../event-system/services/event-observer.interface'
 import { RundownEventParser } from './rundown-event-parser.interface'
 import { RundownEventType } from '../../models/rundown-event-type'
