+++ src/app/core/services/http-ad-lib-piece.service.spec.ts
@@ -1,12 +1,12 @@
-import { AdLibPieceService } from './ad-lib-piece.service';
+import { HttpAdLibPieceService } from './http-ad-lib-piece.service';
 import { HttpClient } from '@angular/common/http'
 import { instance, mock } from '@typestrong/ts-mockito'
 import { HttpErrorService } from './http-error.service'
 
-describe('AdLibPieceService', () => {
+describe('HttpAdLibPieceService', () => {
   it('should be created', () => {
     const mockedHttpClient = mock<HttpClient>()
     const mockedHttpErrorService = mock<HttpErrorService>()
-    expect(new AdLibPieceService(instance(mockedHttpClient), instance(mockedHttpErrorService))).toBeTruthy();
+    expect(new HttpAdLibPieceService(instance(mockedHttpClient), instance(mockedHttpErrorService))).toBeTruthy();
   });
 });
