+++ src/app/core/services/rundown-state.service.spec.ts
@@ -2,7 +2,7 @@ import { RundownStateService } from './rundown-state.service'
 import { instance, mock } from '@typestrong/ts-mockito'
 import { HttpRundownService } from './http-rundown.service'
 import { RundownEventObserver } from './events/rundown-event-observer.service'
-import { ConnectionStatusObserver } from './events/connection-status-observer.service'
+import { ConnectionStatusObserver } from '../../event-system/services/connection-status-observer.service'
 
 describe('RundownStateService', () => {
   it('should be created', () => {
