+++ src/app/rundown/components/rundown/rundown.component.spec.ts
@@ -5,8 +5,8 @@ import { HttpRundownService } from '../../../core/services/http-rundown.service'
 import { anyString, anything, instance, mock, when } from '@typestrong/ts-mockito'
 import { ActivatedRoute, ActivatedRouteSnapshot, convertToParamMap, ParamMap, RouterModule } from '@angular/router'
 import { RundownStateService } from '../../../core/services/rundown-state.service'
-import { AdLibPieceService } from '../../../core/services/ad-lib-piece.service'
-import { of, Subscription, SubscriptionLike } from 'rxjs'
+import { AdLibPieceService } from '../../../core/services/ad-lib-piece-service.interface'
+import { of, Subscription } from 'rxjs'
 
 describe('RundownComponent', () => {
   it('should create', async () => {
